requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
pyserial==3.5
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
